bsl: do not use 2 messages in i2c_write_read
Some I2C bus linux drivers only have support for EEPROM access in random
read. That means writing a message with a higher amount of bytes is not
supported. Since we need to write more than 2 bytes we changed to access
to a separatet write/read message that does not have this limitation.

[1] https://elixir.bootlin.com/linux/v6.6.41/source/drivers/i2c/busses/i2c-at91-master.c#L726

Signed-off-by: Heiko Thiery <heiko.thiery@kontron.com>